Mise à jour PSR-15

+++ src/Framework/Upload.php
@@ -15,24 +15,31 @@ class Upload
         }
     }
 
-    public function upload(UploadedFileInterface $file, ?string $oldFile = null): string
+    /**
+     * @param UploadedFileInterface $file
+     * @param null|string $oldFile
+     * @return null|string
+     */
+    public function upload(UploadedFileInterface $file, ?string $oldFile = null): ?string
     {
-        $this->delete($oldFile);
-        $targetPath = $this->addSuffix($this->path . DIRECTORY_SEPARATOR . $file->getClientFilename());
-        $dirname = pathinfo($targetPath, PATHINFO_DIRNAME);
-        if (!file_exists($dirname)) {
-            mkdir($dirname, 777, true);
+        if ($file->getError() === UPLOAD_ERR_OK) {
+            $this->delete($oldFile);
+            $targetPath = $this->addCopySuffix($this->path . DIRECTORY_SEPARATOR . $file->getClientFilename());
+            $dirname = pathinfo($targetPath, PATHINFO_DIRNAME);
+            if (!file_exists($dirname)) {
+                mkdir($dirname, 777, true);
+            }
+            $file->moveTo($targetPath);
+            $this->generateFormats($targetPath);
+            return pathinfo($targetPath)['basename'];
         }
-        $file->moveTo($targetPath);
-        return pathinfo($targetPath)['basename'];
+        return null;
     }
 
-    private function addSuffix(string $targetPath): string
+    private function addCopySuffix(string $targetPath): string
     {
         if (file_exists($targetPath)) {
-            $info = pathinfo($targetPath);
-            $targetPath = $info['dirname'] . DIRECTORY_SEPARATOR . $info['filename'] . '_copy.' . $info['extension'];
-            return $this->addSuffix($targetPath);
+            return $this->addCopySuffix($this->getPathWithSuffix($targetPath, 'copy'));
         }
         return $targetPath;
     }
@@ -40,7 +47,7 @@ class Upload
     public function delete(?string $oldFile): void
     {
         if ($oldFile) {
-            $oldFile = $this->path . '/' . $oldFile;
+            $oldFile = $this->path . DIRECTORY_SEPARATOR . $oldFile;
             if (file_exists($oldFile)) {
                 unlink($oldFile);
             }
